services/api-py/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
solders
solana
python-dotenv
asyncpg
sqlalchemy[asyncio]
pynacl
base58
py-tlsh
open-clip-torch
torch
torchvision
pgvector
timm
pillow
PyMuPDF
